requirements.txt
asgiref==3.3.1
Django==3.1.7
djangorestframework==3.12.2
Pillow==8.1.2
psycopg2-binary==2.8.6
pytz==2021.1
sqlparse==0.4.1
django-ckeditor==5.3.0
drf-ysg==1.20.0
uritemplate==3.0.1
PyYaml==5.4
